Use a CDN to optimise the delivery of static assets
It assumes the Rails asset pipeline is used to package and serve assets. If not,
whatever solution is used to achieve the same result will need to fingerprint
asset filenames with some sort of unique identifier for it to work smoothly with
Cloudfront or a CDN like it.

NOTE: The config variable is set in the Heroku dashboard.

More info:
https://aws.amazon.com/cloudfront
https://www.happybearsoftware.com/use-cloudfront-and-the-rails-asset-pipeline-to-speed-up-your-app.html
https://devcenter.heroku.com/articles/using-amazon-cloudfront-cdn